更新

+++ source/nanite_builder.h
@@ -7,7 +7,7 @@ namespace Nanite {
 using Meshlet = meshopt_Meshlet;
 
 struct MeshletsContext {
-    std::vector<uint8>   primitives; // a set of indices that define the triangles in the meshlet
+    std::vector<uint32>  primitives; // a set of indices that define the triangles in the meshlet
     std::vector<uint32>  indices; // a list of vertex indices mappings
     std::vector<Meshlet> meshlets;
 };
